backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.7.1
mediapipe==0.10.14
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.3.0
httpx>=0.27
python-dateutil>=2.9
